+++ back/stock/src/main/java/com/ssafy/stock/domain/service/response/StockPricesResponseDto.java
@@ -13,7 +13,7 @@ public class StockPricesResponseDto {
     private String stockIndustry; // 종목 카테고리
     private Long stckPrpr;      // 주식 현재가
     private Long prdyVrss;      // 전일 대비
-    private Long prdyCtrt;      // 전일 대비율
+    private Double prdyCtrt;      // 전일 대비율
     private Long acmlTrPbmn;    // 누적 거래 대금
     private Long acmlVol;       // 누적 거래량
